Begin to add in endpoint for 24hr analytics.
